Move displayHeading to logic and convert to array func
